Entrega 3

+++ web-service/webumvral/web/models/answer.py
@@ -6,4 +6,4 @@ class AnswerModel(models.Model):
     test = models.ForeignKey('TestModel', on_delete=models.CASCADE, related_name='answers')
     student = models.ForeignKey('StudentModel', on_delete=models.CASCADE, related_name='myanswers')
     content = models.CharField(max_length=300, null=True, blank=True)
-    score = models.IntegerField(null=True, blank=True)
+    score = models.FloatField(null=True, blank=True)
